+++ striplog/component.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 Defines components for holding properties of rocks or samples or whatevers.
 
@@ -7,6 +6,7 @@ Defines components for holding properties of rocks or samples or whatevers.
 """
 import json
 
+from .lexicon import Lexicon
 from .utils import CustomFormatter
 
 
@@ -51,6 +51,11 @@ class Component(object):
 
     def __len__(self):
         return len(self.__dict__)
+
+    def __copy__(self):
+        return Component(self.__dict__.copy())
+
+    copy = __copy__
 
     def __iter__(self):
         return iter(self.__dict__)
@@ -143,7 +148,7 @@ class Component(object):
         return json.dumps(self.__dict__)
 
     @classmethod
-    def from_text(cls, text, lexicon, required=None, first_only=True):
+    def from_text(cls, text, lexicon=None, required=None, first_only=True):
         """
         Generate a Component from a text string, using a Lexicon.
 
@@ -160,11 +165,32 @@ class Component(object):
             Component: A Component object, or None if there was no
                 must-have field.
         """
+        if lexicon is None:
+            lexicon = Lexicon.default()
         component = lexicon.get_component(text, first_only=first_only)
         if required and (required not in component):
             return None
         else:
             return cls(component)
+
+    @classmethod
+    def from_macrostrat(cls, feature, columns=None):
+        """
+        Make a Component from a Macrostrat feature dictionary.
+        """
+        props = feature['properties']
+        if columns is None:
+            columns = []
+        for column in columns:
+            if props.get(column) is None:
+                if column == 'color': # skip if no colour
+                    continue
+                props[column] = '' # Need a value to match on.
+                continue
+            else:
+                props[column] = props[column].replace(',', ';').lower()
+        props['source'] = 'Macrostrat.org (CC-BY)'
+        return cls(props)
 
     def summary(self, fmt=None, initial=True, default=''):
         """