+++ striplog/position.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 Defines positions, eg for top and base of an interval.