+++ striplog/rock.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 Replaced by component.py. Kept this for graceful deprecation.