+++ striplog/striplog.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 A striplog is a sequence of intervals.
 
@@ -18,6 +17,8 @@ from copy import deepcopy
 import numpy as np
 import matplotlib as mpl
 import matplotlib.pyplot as plt
+import requests
+import json
 
 from .interval import Interval, IntervalError
 from .component import Component
@@ -678,7 +679,7 @@ class Striplog(object):
         return cls(list_of_Intervals, source=source)
 
     @classmethod
-    def from_dict(cls, dict,
+    def from_dict2(cls, dict,
                  lexicon=None,
                  points=False,
                  include=None,
@@ -691,7 +692,7 @@ class Striplog(object):
                  stop=None,
                  ):
         """
-        Load from a CSV file or text.
+        Load from a dictionary reusing `from_csv` post processing.
 
         Args
             dict (dict): python dict containing the data.
@@ -732,6 +733,25 @@ class Striplog(object):
                                                          stop=stop)
 
         return cls(list_of_Intervals, source=source)
+
+    @classmethod
+    def from_dict(cls, dictionary):
+        """
+        Take a dictionary of the form name:depth and return a striplog of
+        complete intervals.
+        """
+        d_sorted = sorted(dictionary.items(), key=lambda i: i[1])
+        names = [i[0] for i in d_sorted]
+        tops_ = [i[1] for i in d_sorted]
+        bases_ = tops_[1:] + [tops_[-1]+1]
+        comps_ = [Component({'formation': name}) for name in names]
+
+        list_of_Intervals = []
+        for top, base, comp in zip(tops_, bases_, comps_):
+            iv = Interval(top=top, base=base, components=[comp])
+            list_of_Intervals.append(iv)
+
+        return cls(list_of_Intervals)
 
     @classmethod
     def from_descriptions(cls, text,
@@ -911,6 +931,9 @@ class Striplog(object):
                                                       values,
                                                       basis,
                                                       components)
+
+        list_of_Intervals = [iv for iv in list_of_Intervals
+                             if isinstance(iv.primary, Component)]
 
         return cls(list_of_Intervals, source="Image")
 
@@ -1128,6 +1151,10 @@ class Striplog(object):
                         order=self.order,
                         source=self.source)
 
+
+
+
+
     # Outputter
     def to_canstrat(self, filename, params):
         """
@@ -1231,8 +1258,10 @@ class Striplog(object):
                basis=None,
                field=None,
                field_function=None,
-               dtype=None,
+               bins=True,
+               dtype='float',
                table=None,
+               sort_table=False,
                legend=None,
                legend_field=None,
                match_only=None,
@@ -1256,6 +1285,12 @@ class Striplog(object):
             field_function (function): Provide a function to apply to the field
                 you are asking for. It's up to you to make sure the function
                 does what you want.
+            bins (bool): Whether to return the index of the items from the
+                lookup table. If False, then the item itself will be returned.
+            dtype (str): The NumPy dtype string for the output log.
+            table (list): Provide a look-up table of values if you want. If you
+                don't, then it will be constructed from the data.
+            sort_table (bool): Whether to sort the table or not. Default: False.
             legend (Legend): If you want the codes to come from a legend,
                 provide one. Otherwise the codes come from the log, using
                 integers in the order of prevalence. If you use a legend,
@@ -1331,6 +1366,9 @@ class Striplog(object):
         else:
             match_only = []
 
+        if sort_table:
+            table.sort()
+
         start_ix = self.read_at(start, index=True)
         stop_ix = self.read_at(stop, index=True)
         if stop_ix is not None:
@@ -1353,7 +1391,12 @@ class Striplog(object):
                 f = field_function or utils.null
                 try:
                     v = f(i.data.get(field, undefined)) or undefined
-                    key = (table.index(v) + 1) or undefined
+                    if bins:
+                        # Then return the bin we're in...
+                        key = (table.index(v) + 1) or undefined
+                    else:
+                        # ...else return the actual value.
+                        key = v
                 except ValueError:
                     key = undefined
             else:  # Use the lookup table.
@@ -1759,8 +1802,9 @@ class Striplog(object):
             function (function). A function that takes an array as the only
                 input, and returns whatever you want to store in the 'name'
                 attribute of the primary component.
+
         Returns:
-            None. The function works on the striplog in place.
+            A copy of the striplog.
         """
         # Build a dict of {index: [log values]} to keep track.
         intervals = {}
@@ -1776,12 +1820,13 @@ class Striplog(object):
             previous_ix = ix
 
         # Set the requested attribute in the primary comp of each interval.
+        new_strip = self.copy()
         for ix, data in intervals.items():
             f = function or utils.null
             d = f(np.array(data))
-            self[ix].data[name] = d
+            new_strip[ix].data[name] = d
 
-        return None
+        return new_strip
 
     def find(self, search_term, index=False):
         """
@@ -1797,6 +1842,7 @@ class Striplog(object):
             search_term (string or Component): The thing you want to search
                 for. Strings are treated as regular expressions.
             index (bool): Whether to return the index instead of the interval.
+
         Returns:
             Striplog: A striplog that contains only the 'hit' Intervals.
                 However, if ``index`` was ``True``, then that's what you get.
@@ -2625,3 +2671,78 @@ class Striplog(object):
             proc = proc | log
 
         return Striplog.from_log(proc, components=comps, basis=basis)
+
+    @classmethod
+    def from_macrostrat(cls, lng, lat, buffer_size=0.2):
+        """
+        Create a striplog from components derived using the MacroStrat API.
+            This is simply a helper function to make things easier, but it
+            works because we know what our data looks like in advance.
+
+        Note: In order to plot this, you will need to add space for text and
+            other decoration. This simply gives a Striplog back which _can_
+            be plotted.
+
+        Args:
+            components (list):
+
+        Returns:
+            Tuple of:
+                strip (striplog.Striplog)
+                legend (striplog.Legend)
+
+        Example:
+        lng = -64.3573186
+        lat = 44.4454632
+        buffer_size = 0.3
+        striplog.striplog.from_macrostrat(lng, lat, buffer_size)
+        {'top': Position({'middle': 358.9, 'units': 'm'}),
+            'base': Position({'middle': 419.2, 'units': 'm'}),
+            'description': '', 'data': {}, 'components': [Component({
+                'map_id': 948660.0, 'scale': 'small', 'source_id': 7.0,
+                'name': 'Devonian plutonic: undivided granitic rocks',
+                'age': 'devonian', 'lith': 'plutonic: undivided granitic rocks',
+                'best_age_top': 358.9, 'best_age_bottom': 419.2, 't_int': 94.0,
+                'b_int': 94.0, 'color': '#cb8c37', 'source': 'MacroStrat.org (CC-BY)})]}
+        {'top': Position({'middle': 358.9, 'units': 'm'}),
+            'base': Position({'middle': 541.0, 'units': 'm'}),
+            'description': '', 'data': {}, 'components': [Component({
+                'map_id': 948228.0, 'scale': 'small', 'source_id': 7.0,
+                'name': 'Cambrian-Devonian sedimentary', 'age': 'cambrian-devonian',
+                'lith': 'sedimentary', 'best_age_top': 358.9, 'best_age_bottom': 541.0,
+                't_int': 94.0, 'b_int': 122.0, 'color': '#99c08d',
+                'source': 'MacroStrat.org (CC-BY)})]}
+        {'top': Position({'middle': 443.8, 'units': 'm'}),
+            'base': Position({'middle': 541.0, 'units': 'm'}),
+            'description': '', 'data': {}, 'components': [Component({
+                'map_id': 973359.0, 'scale': 'small', 'source_id': 7.0,
+                'name': 'Cambrian-Ordovician sedimentary', 'age': 'cambrian-ordovician',
+                'lith': 'sedimentary', 'best_age_top': 443.8, 'best_age_bottom': 541.0,
+                't_int': 112.0, 'b_int': 122.0, 'color': '#409963',
+                'source': 'MacroStrat.org (CC-BY)})]}
+        """
+        # Get the
+        features = utils.geology_from_macrostrat(lng=lng, lat=lat,
+                                                 buffer_size=buffer_size)
+
+        columns = ('color', 'lith', 'age')
+
+        intervals = []
+
+        for feature in features:
+            if feature['geometry'] is None:
+                continue
+
+            components = []
+            for lith in utils.get_liths_from_macrostrat(feature['properties']['lith']):
+                c = Component({'lithology': lith})
+                components.append(c)
+
+            intervals.append(Interval(
+                            top=feature['properties']['best_age_top'],
+                            base=feature['properties']['best_age_bottom'],
+                            components=components,
+                            description=feature['properties']['descrip'])
+                            )
+
+        return cls(intervals, source='Macrostrat [CC-BY]', order='age')