+++ tests/test_lexicon.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 Define a suite a tests for the Lexicon module.
 """
@@ -13,7 +12,7 @@ def test_lexicon():
     assert s is not ''
     assert lexicon.__repr__() is not ''
     assert lexicon.find_synonym('Halite') == 'salt'
-    assert len(lexicon.categories) == 4
+    assert len(lexicon.categories) == 5
 
     s = "lt gn ss w/ sp gy sh"
     answer = 'lighter green sandstone with spotty gray shale'