+++ tests/test_plots.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 Defines a suite a tests for images run with: