+++ CHANGES.rst
@@ -1,6 +1,35 @@
 Changelog
 =========
 
+0.8.8 (January 2021)
+----
+
+- Added `Striplog.from_dict()` to create a striplog from a dictionary with items like `name: depth`.
+- When reading a Petrel 'tops' file, you can pass a single function to apply to every quantity, not just a mapping of field names to functions. (You might do this to replace all `-999.25` with `np.nan`, for example.) There's no way to do both at present, but if there's a need, please file an issue.
+- Updated the default lexicon with some 'modifiers' like 'silty', 'sandy', etc.
+- `Lexicon._parse_description()` is now a private static method (it was previously mangled which was a mistake). It can be called without instantiating a class to parse a piece of text. See the notebook, `Parsing_a_description.ipynb`.
+- Fixed a bug causing the default Lexicon to be a shallow copy, which was annoying.
+
+
+0.8.7 (November 2020)
+----
+
+- Breaking change: `Striplog.extract()` now returns a copy of the striplog instead of changing the striplog in place.
+- Added option to send actual data to `Striplog.to_log()` instead of index from lookup table.
+- Added a tutorial, `Blocking_curves_with_striplog.ipynb`.
+
+
+0.8.6 (October 2020)
+----
+
+- Multi-step Markov chains now work properly, respecting 'self-transitions'.
+- Added `include_self` property to the repr of a Markov chain object.
+- Added `match_only` argument to `Legend.random()` to allow you to keep only certain properties in the components.
+- `Component.from_text()` will now use `Lexicon.default()` if you don't pass one in.
+
+We started adding an experimental new feature; feedback welcome! Added `Striplog.from_macrostrat()` which takes a lat/lon location and returns a striplog in geological time or in depth below the surface at that location.
+
+
 0.8.5 (May 2020)
 ----